Move translate_annotation_principals to app
