+++ h/api/test/auth_test.py
@@ -4,36 +4,10 @@ import datetime
 
 from jwt import InvalidTokenError
 import mock
-from pyramid import security
-from pyramid import testing
 import pytest
 
 from h.api import auth
 from h.api.models.token import API_TOKEN_PREFIX
-
-
-@pytest.mark.parametrize("p_in,p_out", [
-    # The basics
-    ([], []),
-    (['acct:donna@example.com'], ['acct:donna@example.com']),
-    (['group:foo'], ['group:foo']),
-
-    # Remove pyramid principals
-    (['system.Everyone'], []),
-
-    # Remap annotatator principal names
-    (['group:__world__'], [security.Everyone]),
-    (['group:__authenticated__'], [security.Authenticated]),
-
-    # Normalise multiple principals
-    (['me', 'myself', 'me', 'group:__world__', 'group:foo', 'system.Admins'],
-     ['me', 'myself', security.Everyone, 'group:foo']),
-])
-def test_translate_annotation_principals(p_in, p_out):
-    result = auth.translate_annotation_principals(p_in)
-
-    assert set(result) == set(p_out)
-
 
 generate_jwt_fixtures = pytest.mark.usefixtures('jwt')
