+++ h/streamer/nsq.py
@@ -10,8 +10,8 @@ import struct
 from gevent.queue import Full
 
 from h import queue
-from h.api import auth
 from h.api import storage
+from h.auth.util import translate_annotation_principals
 from h.streamer import websocket
 
 log = logging.getLogger(__name__)
@@ -175,7 +175,7 @@ def _authorized_to_read(request, annotation):
             })
 
     read_permissions = annotation.get('permissions', {}).get('read', [])
-    read_principals = auth.translate_annotation_principals(read_permissions)
+    read_principals = translate_annotation_principals(read_permissions)
     if set(read_principals).intersection(request.effective_principals):
         return True
     return False
